requirements.txt
fastapi
uvicorn[standard]
duckdb
pandas
bcrypt
python-jose
